backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
click==8.3.1
fastapi==0.135.2
google-auth==2.41.1
h11==0.16.0
httptools==0.7.1
idna==3.11
numpy==2.1.3
pandas==2.2.3
PyJWT==2.10.1
pydantic==2.12.5
pydantic_core==2.41.5
PyMuPDF==1.25.5
python-docx==1.1.2
requests==2.32.5
scikit-learn==1.7.2
python-dotenv==1.2.2
PyYAML==6.0.3
python-multipart==0.0.20
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
